Implement initial Golang backend
This commit adds a rough implementation of a backend in Golang.
This commit also modifies the electron app frontend to allow it to work
with the new backend.

Currently, the frontend communicates with the backend exclusively via
HTTP requests. In the future, it may make sense to request 'suggestions'
via a web socket connection. The backend also may need to be swapped out
with a different solution in the future to improve efficiency.

The backend will handle all Spotify API requests and requests to other
external services if necessary. It currently is made up of just a HTTP
server but in the near future it will include functionality for
caching Spotify user libraries.