Implement Spotify auth
This commit implements a system to authenticate with Spotify
through the built-in UI as well as management of access tokens
and automatic token saving and refreshing.

Initial tests have been written in the `authenticate_test.go`
file to test command suggestions and execution. These kind of
tests should be written for each new command going forward.