Implement Auth command
This commit implements a user-Spotify authentication command in
the frontend of the application. The backend hosts a http server
which will start when the user initiates the authentication process
with Spotify and will finish when the user successfully authenticates
or closes the 'Authenticate' command.

This commit also:
- Adds an optional onCancel method side-effect to the Command interface,
which triggers when a command is removed from the command stack in state.
- Updates and adds new icons
- Updated volume command to show current volume for command placeholder
- Refactors some key listeners from the Spotlightify component to the
useCommand hook.