Add Liked Songs command
This commit adds a 'Liked Songs' command which will play the first 50
saved songs for a user. When retrieving a user's saved songs the maximum
that can be retrieved is 50 per request therefore I decided only one
request should be made to retrieve these songs. This is something that
could possibly be enhanced if/when background caching is implemented and
we can store all of the users saved songs locally so we do not have to
retrieve them from Spotify each time this command is executed.