Fix Spotify API changes issue
This commit fixes an issue that was introduced by changes to Spotify's
API functionality. The issue was that the API does not retrieve
playlists generated by Spotify, so instead of retrieving the data about
these playlists, it would return blank playlists instead. The fact that
the API is returning blank playlists is probably a bug on Spotify's end
and I would it expect it to be fixed quickly anyway.